requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.2.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
waitress==2.0.0
Werkzeug==2.0.2
